// types.ts
export interface BookingRecord {
  fechaVuelo: string;
  fechaHoraCorreo: string;
  estatus: string;
  codigoReserva: string;
  genero: string;
  nombreCompleto: string;
  fechaNacimiento: string;
  edad: number;
  pasaporte: string;
  nacionalidad: string;
  agencia: string;
  fechaInicio: string;
  fechaFin: string;
  noches: number;
  hotel: string;
  planDeComidas: string;
  alojamiento: string;
  observaciones: string;
}